recordhelper/requirements.txt
line-bot-sdk
openai
google-generativeai
Flask
python-dotenv
requests
pydub 